Moved compile code from gulpfiles.js to tools/compile.js.
